move `MinerGetBaseInfo` to miner module
